requirements.txt
streamlit
pandas
numpy
plotly
websocket-client
statsmodels
scipy
